На проверку 2

+++ functions/validate_bet.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * функция валидации ставки
- * @param array $bet_field данные полученные из формы
+ * @param string $bet_field данные полученные из формы
  * @param array $lot данные полученные из базы
  * @return array массив ошибок
  */
@@ -21,11 +21,10 @@ function validate_bet($bet_field, $lot) {
 }
 /**
  * функция валидации суммы ставки
- * @param array $bet_field Ставка полученная из формы
+ * @param string $bet_field Ставка полученная из формы
  * @return string описание ошибки
  */
 function validate_bet_field($bet_field) {
-
   if($bet_field === ''){
     return 'Заполните размер ставки!';
   }
@@ -37,11 +36,11 @@ function validate_bet_field($bet_field) {
 /**
  * функция валидации суммы ставки
  * @param array $lot последняя ставка полученная из базы
- * @param array $bet_field Ставка полученная из формы
+ * @param string $bet_field Ставка полученная из формы
  * @return string описание ошибки
  */
 function validate_bet_size($lot, $bet_field) {
-  $bet = get_value($lot, 'last_bet_amount') + get_value($lot, 'step');
+  $bet = (int)get_value($lot, 'last_bet_amount') + (int)get_value($lot, 'step');
   if($bet_field < $bet){
     return 'Минимальная ставка ' . $bet .'p.' ;
   }
@@ -49,9 +48,9 @@ function validate_bet_size($lot, $bet_field) {
 }
 /**
  * функция валидации формы ставки
- * @param array $lot ставки пользователя
- * @param array $last_bet_user_id ставка пользователя
  * @param array $user Ставка полученная из формы
+ * @param array $lot ставки пользователя
+ * @param string $last_bet_user_id ставка пользователя
  * @return string описание ошибки
  */
 function restrictions($user, $lot, $last_bet_user_id) {
